Quick fix workspace layout.
