+++ client/src/App.jsx
@@ -1,12 +1,12 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import SignUpPage from './components/pages/SignUpPage';
-import SignInPage from './components/pages/SignInPage';
-import WalletsPage from './components/pages/WalletsPage';
-import Transaction from './components/Layout/Transaction';
-import TransactionPage from './components/pages/TransactionsPage';
-import WalletCard from './components/Layout/WalletCard';
-import HistoryCard from './components/Layout/HistoryCard';
-import AddWallet from './components/Layout/AddWallet';
+import SignUpPage from './components/pages/SignUpPage/SignUpPage';
+import SignInPage from './components/pages/SignInPage/SignInPage';
+import WalletsPage from './components/pages/WalletsPage/WalletsPage';
+import Transaction from './components/Layout/AddTransactionComponent/Transaction';
+import TransactionPage from './components/pages/TransactionPage/TransactionsPage';
+import WalletCard from './components/Layout/WalletCard/WalletCard';
+import HistoryCard from './components/Layout/HistoryCard/HistoryCard';
+import AddWallet from './components/Layout/AddWallet/AddWallet';
 import './App.css';
 
 function App() {
@@ -18,12 +18,8 @@ function App() {
         <Routes>
           <Route path='/' element={<SignInPage />} />
           <Route path='/signup' element={<SignUpPage />} />
-          <Route path='/trans' element={<Transaction />} />
-          <Route path='/transaction' element={<TransactionPage />} />
-          <Route path='/card' element={<WalletCard />} />
-          <Route path='/history' element={<HistoryCard />} />
+          <Route path='/transactions' element={<TransactionPage />} />
           <Route path='/wallets' element={<WalletsPage />} />
-          <Route path='/add' element={<AddWallet />} />
         </Routes>
       </Router>
     </>
