+++ client/src/components/Layout/AddTransactionComponent/Transaction.jsx
@@ -10,7 +10,7 @@ import MenuItem from '@mui/material/MenuItem';
 import InputAdornment from '@mui/material/InputAdornment';
 import InputLabel from '@mui/material/InputLabel';
 import { Input, Label, FormGroup, FormText } from 'reactstrap';
-import '../../styles/TransactionComponent.css';
+import './TransactionComponent.css';
 
 const ITEM_HEIGHT = 48;
 const ITEM_PADDING_TOP = 8;
