+++ client/src/components/Layout/HistoryCard/HistoryCard.jsx
@@ -5,7 +5,7 @@ import CardActions from '@mui/material/CardActions';
 import CardContent from '@mui/material/CardContent';
 import Button from '@mui/material/Button';
 import Typography from '@mui/material/Typography';
-import '../../styles/HistoryCard.css';
+import './HistoryCard.css';
 
 const HistoryCard = ({name, amount}) => {
   const cardStyle = {
