+++ client/src/components/Layout/TransactionCard/TransactionCard.css
@@ -8,7 +8,7 @@
   left: 45%;
   top: 80%;
   transform: translate(-50%, -50%);
-  background-color: white;
+  background-color: transparent;
 }
 
 .card-media {
@@ -19,13 +19,13 @@
   left: 45%;
   top: 40%;
   transform: translate(-50%, -50%);
+  background-color: transparent;
 }
 
 .full-card {
   align-content: center;
   box-shadow: none;
   background-color: transparent;
-  width: 250px;
   height: 150px;
 }
